Complete modular demuxer/codec architecture implementation
- Implement comprehensive demuxer architecture with RIFF, Ogg, and ISO container support
- Add RIFFDemuxer for WAV files with PCM/A-law/μ-law/MP3 codec support
- Create RawAudioDemuxer for containerless formats (.ulaw, .alaw, .pcm, etc.)
- Implement OggDemuxer for Vorbis/FLAC/Opus/Speex container parsing
- Add ISODemuxer for MP4/M4A file format with AAC/ALAC/MP3 codec identification
- Integrate all demuxers into DemuxerFactory with auto-detection
- Centralize all includes in psymp3.h following unity build approach
- Update build system (Makefile.am, psymp3.final.cpp) with new components
- Add comprehensive codec framework with PCM variants and telephony formats
- Create DemuxedStream bridge for backwards compatibility with existing Stream interface

The modular architecture now supports:
- RIFF containers: WAV with multiple PCM formats and codec passthrough
- Raw audio: Extension-based detection for telephony and broadcast formats
- Ogg containers: Multi-codec support with proper page/packet parsing
- ISO containers: MP4/M4A with sample table interpretation and track management

OggCodecs contains placeholder implementations for future codec passthrough development.

+++ src/RIFFDemuxer.cpp
@@ -7,9 +7,7 @@
  * the terms of the ISC License <https://opensource.org/licenses/ISC>
  */
 
-#include "RIFFDemuxer.h"
-#include "exceptions.h"
-#include <algorithm>
+#include "psymp3.h"
 
 RIFFDemuxer::RIFFDemuxer(std::unique_ptr<IOHandler> handler) 
     : Demuxer(std::move(handler)) {
@@ -220,7 +218,7 @@ bool RIFFDemuxer::parseWaveFormat(const RIFFChunk& chunk) {
     // Read extra data if present
     if (chunk.size > 16) {
         uint16_t extra_size = readLE<uint16_t>();
-        if (extra_size > 0 && chunk.size >= 18 + extra_size) {
+        if (extra_size > 0 && chunk.size >= 18 + static_cast<uint32_t>(extra_size)) {
             stream_data.extra_data.resize(extra_size);
             m_handler->read(stream_data.extra_data.data(), 1, extra_size);
         }
